data_analyze_server/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
asyncpg
psycopg2-binary        # 개발용
python-dotenv
paho-mqtt
pytest
httpx   # fastapi[test]에 포함됨
pydantic-settings
datetime
typing
apscheduler
scikit-learn
pandas
xgboost
redis